requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
certifi==2024.7.4
cffi==1.17.1
cfgv==3.4.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
diagrams==0.24.4
distlib==0.3.9
Django==4.2.15
django-auditlog==3.0.0
django-cors-headers==4.7.0
django-jsonfield==1.4.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-reversion==5.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
et-xmlfile==1.1.0
filelock==3.18.0
graphviz==0.20.3
hyperlink==21.0.0
identify==2.6.9
idna==3.7
importlib-metadata==6.7.0
incremental==24.7.2
inflection==0.5.1
isodate==0.6.1
jdcal==1.4.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.0.5
mysqlclient==2.2.4
nodeenv==1.9.1
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pre_commit==4.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.36.2
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
reversion==0.2
rpds-py==0.23.1
service-identity==24.2.0
setuptools==77.0.3
six==1.16.0
sqlparse==0.5.1
toml==0.10.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.29.3
XlsxWriter==3.2.0
zeep==4.2.1
zipp==3.15.0
zope.interface==7.2
gunicorn==21.2.0
